AES methods reorder as previously discussed
