+++ Transformations/AESmethods.cpp
@@ -1,3 +1,5 @@
+#include "AESmethods.h"
+
 // Predetermined s-box for Rjindael
 unsigned char sBox[16][16] = {
     {0x63, 0x7c, 0x77, 0x7b, 0xf2, 0x6b, 0x6f, 0xc5, 0x30, 0x01, 0x67, 0x2b, 0xfe, 0xd7, 0xab, 0x76},
